+++ src/models/PriceAdjustment.js
@@ -5,10 +5,10 @@ const PriceAdjustment = async () => {
     if (!await db.schema.hasTable('priceAdjustments')){
 
         await db.schema.createTable('priceAdjustments', table => {
-            table.bigIncrements('id').primary();
-            table.bigInteger('userId').notNullable();
+            table.increments('id').primary();
+            table.integer('userId').notNullable();
             table.date('date').defaultTo(db.fn.now());
-            table.bigInteger('productId');
+            table.integer('productId');
             table.float('oldCost');
             table.float('newCost');
             table.float('oldPrice');
