+++ src/models/SalesReturn.js
@@ -4,9 +4,9 @@ const SalesReturn = async () => {
     if (!await db.schema.hasTable('salesReturns')){
 
         await db.schema.createTable('salesReturns', table => {
-            table.bigIncrements('id').primary();
-            table.bigInteger('userId').unsigned();
-            table.bigInteger('productId').unsigned();
+            table.increments('id').primary();
+            table.integer('userId').unsigned();
+            table.integer('productId').unsigned();
             table.string('productName')
             table.integer('quantity', 7);
             table.float('buyingPrice');
@@ -16,8 +16,8 @@ const SalesReturn = async () => {
             table.float('tax');
             table.float('discount');
             table.date('date').defaultTo(db.fn.now());
-            table.bigInteger('categoryId').unsigned();
-            table.bigInteger('orderId').unsigned().notNullable();
+            table.integer('categoryId').unsigned();
+            table.integer('orderId').unsigned().notNullable();
 
             table.foreign('orderId').references('id').inTable('orders').onDelete('CASCADE');
             table.foreign('categoryId').references('id').inTable('categories');
